Moving common tasks to a different package in the same project
