+++ sisifo_common/parallel_pipeline.py
