+++ sisifo_common/pipeline.py
