--- tests/__init__.py
@@ -1,5 +1,0 @@
-import pytest
-
-
-# we want to have pytest assert introspection in the sisifo package
-pytest.register_assert_rewrite("sisifo")
