+++ tests/unit/utils/test_named_collection.py
@@ -1,4 +1,4 @@
-import sisifo.utils.named_collection as named_collection
+import sisifo.named_collection as named_collection
 
 import pytest
